Is this what it means to have a singleton?
- I learned that user defined classes can’t have multiple inheritance
despite being able to outside of Unreal Engine, so my plan of calling a
draw function in my Item function won’t work. Moving
everything that will be needed to render text when item is picked up to
it’s own ItemTextHUD class.

- Despite having multiple hints telling me that HUD classes can only be
static, I’ve realized that there can only be one HUD per game mode. I
created an ItemTextHUD class and created functions around the assumption
that I would be able to have an instance of it in the Item class, so
I’ll be writing a general HUD class that includes the ItemTextHUD class
and goes through each item instance in the level to check if one has
been picked up or attempted to be picked up.