Listen to that little voice in your head that tells you to stop going down the rabbit hole.
-added some basic descriptions on some headers so even I don't forget
what they're for.

-Spent probably more than half the day trying to get a text widget
instance in the item class to work. I'm not sure what exactly the
underlying issue was but it took too much time so I went with the HUD
implementation

-Implemented HUD text for Item but I wasn't seeing it again and spent a
good amount of time trying to figure out why that was by having static
coordinates. Realized translation between 3D coordinates from the item
and 2D coordinates for the text wasn't happening. Investigated if there
was a built-in function for this, and there is specifcally for HUD
things. Going to work on scaling text depending how close or far the
camera is