The easy way to detect when player is outside of item's collision range
- I learned that a variable that is of type "someSignature" can be
invocated like any object with '.', there is no need to add '()' then
'.'
-I added the "OverlapEnd" function to make it easy to detect when the
player is out of range to pickup an item
-I removed lines in the "switch collision" function because setting
overlap bool variables to false crashes UE, because the OverlapEnd is
not able to work. (will removing these lines of code break functionality
for when an item actually needs to be picked up?
-TODO: create an additional array in ItemTextHUD for all the fail text
and pair with reference/address to the current item that called it.