Item Class Fully Implemented (pre-inventory implementation)
-Derived UIText class into the success and fail subclasses so they can
contain their specific member variable (fail needs ptr to the item it's
displaying text for while success only needs the location of the item)
-I had to create a "GetClassType" virtual function for those classes
because dynamic_cast doesn't work in UE4, Cast function only works for
UObjects. Thankfully this was not much work with the help of enums
-I made the points at which the object location switches between in a
fail case closer together so the animation doesn't look to crazy
-I created a boolean to check if the display text for an item is set or
not so it doesn't display more than 1 text at a time.
-I moved the display text drawing logic to its own function so it's not
cluttering the Draw function