Oh how useful reading source header files is!
-Probably saved a lot of time looking through source files to figure out
what is meant by : void ForEachWidget(Predicate Pred) ....turns out that
means that you can use a lambda function!
-Had the darndest time trying to figure out in my simple inventory why
my fully flegged widget wasn't showing any of its components. Turns out
that having widget components be children of a canvas means that the
positioning is absolute. I created a copy of my original widget and
recreating it without the use of canvas
-Need to dynamically size fonts for the text that show the name of the
item and the amount of it, so it fits in whatever I believe to be a good
size for the slot in the end.
-Currently trying to go through child widgets of my slotwidgets and
determine the correct text that will need to be resized, and how much
they will need to be resized.