Get it working, then optimize where needed
-Re-numbered the item slots and horizontal slots in the inventory so
that the first one is the leftmost slot, and the second to bottom slot,
respectively. Since the currently highlighted one will always be the
seciond to bottom slot and and leftmost slot.

-Added FLinearColor fields in SlotWidget.h for cases when the slot is
highlighted, in the same row as a highlighted slot, or when it's
passive/not in focus.

-Added getters and setters for the fields mentioned above

-Attempted to make the FLinearColor fields static but the compiler
complained that it was an unrecognized type.

-In progress of setting defaults for colors in slotwidget class

TODO: Rework InventoryWidget.cpp so that it initializes what is
currently in the UMG view but hides all but one horizontal section and 1
slot and indicator