Inventory only appears if being called through Player Controller?
-I realized that I had code that I got from a tutorial in my character
class and thought this is what was causing the inventory to appear on
screen. It turns out that it's not responding to this code, but to code
that's been added to the player controller.... Why only from that class
and not from both?

-Not sure of an ideal solution for where widget initialize code should
live if it's not under the Initialize function. Currently get errors
before gameplay has begun saying that widget couldn't be created because
the world doesn't exist. But I'm not calling it before play....