Many a fixes
-added temp 2d texture
-moved code in ctors of indicator and slot widget into init functions
-Added condirional to check if world exists before attempting to go
through init code in horizwidget, indicator widget and slot widget.
-Added init of super class in slot and indicator classes otherwise
attempting to open their BP in the editor would result in crash.
-calling inventorywidget show and hide functions weren't working because
the entire widget visibility was hidden, so I changed it to toggle the
visible of the whole widget for now
-changed 2dtexture references to be editdefaultsonly in BP since
createdefaultsubobject doesn't work with them
-Realized that part of why the inventory wasn't showing changes is
because the name the widget tree was trying to find doesn't exist, so
that's fixed
-Realized that Inventory init function is only setting 1 horizontal slot
from the BP. TODO: ensure it's changing all horizontal widgets in
inventory BP.